requirements.txt
discord.py
requests
python-decouple
python-dotenv
pynacl
unidecode
mysql-connector-python
easy_pil
num2words
PyGitHub
pyMorseTranslator
caesarcipher
googletrans==4.0.0rc1
Faker